Do the same for reading
